# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ambassador Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sync/trigger:
    post:
      operationId: triggerSync
      tags: [sync]
      summary: Trigger manual sync
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerSyncRequest"
      responses:
        "200":
          description: Sync triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJobResponse"
  /sync/jobs:
    get:
      operationId: listSyncJobs
      tags: [sync]
      summary: List sync jobs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Sync job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJobListResponse"
  /sync/jobs/{id}:
    get:
      operationId: getSyncJob
      tags: [sync]
      summary: Get sync job details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sync job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJobResponse"
...
components:
  schemas:
...
    TriggerSyncRequest:
      type: object
      properties:
        jobType:
          type: string
          enum: [FULL, AMBASSADORS_ONLY, REFERRALS_ONLY]
          default: FULL
    SyncJob:
      type: object
      properties:
        id:
          type: integer
        jobType:
          type: string
        status:
          type: string
        recordsProcessed:
          type: integer
        recordsFailed:
          type: integer
        errorLog:
          type: string
          nullable: true
        startedAt:
          type: string
        completedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - jobType
        - status
        - recordsProcessed
        - recordsFailed
        - startedAt
        - createdAt
    SyncJobResponse:
      type: object
      properties:
        job:
          $ref: "#/components/schemas/SyncJob"
      required:
        - job
    SyncJobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/SyncJob"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - jobs
        - total
        - page
        - limit
...