# lib/api-zod/src/generated/types/syncJobListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ambassador Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { SyncJob } from "./syncJob";

export interface SyncJobListResponse {
  jobs: SyncJob[];
  total: number;
  page: number;
  limit: number;
}